Staff Modifications Done

+++ backend/src/main/java/com/pos/staff/dao/OrderItemDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.pos.staff.entity.OrderItem;
 
-public interface OrderItemDao extends JpaRepository<OrderItem,Long>{
-
+public interface OrderItemDao extends JpaRepository<OrderItem,Integer>{
+	
 }
